+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -268,5 +268,9 @@ export default {
     'es.iterator.take',
     'es.iterator.to-array',
     'es.promise.try',
+    'esnext.map.get-or-insert',
+    'esnext.map.get-or-insert-computed',
+    'esnext.weak-map.get-or-insert',
+    'esnext.weak-map.get-or-insert-computed',
   ],
 };