+++ packages/core-js/full/map/index.js
@@ -2,6 +2,7 @@
 var parent = require('../../actual/map');
 require('../../modules/esnext.map.from');
 require('../../modules/esnext.map.of');
+require('../../modules/esnext.map.key-by');
 require('../../modules/esnext.map.delete-all');
 require('../../modules/esnext.map.emplace');
 require('../../modules/esnext.map.every');
@@ -9,7 +10,8 @@ require('../../modules/esnext.map.filter');
 require('../../modules/esnext.map.find');
 require('../../modules/esnext.map.find-key');
 require('../../modules/esnext.map.includes');
-require('../../modules/esnext.map.key-by');
+require('../../modules/esnext.map.get-or-insert');
+require('../../modules/esnext.map.get-or-insert-computed');
 require('../../modules/esnext.map.key-of');
 require('../../modules/esnext.map.map-keys');
 require('../../modules/esnext.map.map-values');