+++ packages/core-js/full/weak-map/from.js
@@ -6,6 +6,8 @@ require('../../modules/es.weak-map');
 require('../../modules/esnext.weak-map.from');
 require('../../modules/esnext.weak-map.delete-all');
 require('../../modules/esnext.weak-map.emplace');
+require('../../modules/esnext.weak-map.get-or-insert');
+require('../../modules/esnext.weak-map.get-or-insert-computed');
 require('../../modules/web.dom-collections.iterator');
 var path = require('../../internals/path');